move calibration check modules to calibration package
