+++ api/src/opentrons/server/endpoints/calibration/check.py
@@ -1,8 +1,8 @@
 from aiohttp import web
 
 from opentrons import types
-from .session import CheckCalibrationSession, CalibrationCheckTrigger
-from .models import JogPosition
+from opentrons.calibration.check.session import CheckCalibrationSession, CalibrationCheckTrigger
+from opentrons.calibration.check.models import JogPosition
 
 
 async def get_session(request: web.Request, session) -> web.Response:
