+++ robot-server/robot_server/service/session/session_status.py
@@ -1,7 +1,7 @@
 import typing
-from opentrons.server.endpoints.calibration import models as calibration_models
-from opentrons.server.endpoints.calibration.session import \
-    CheckCalibrationSession, CalibrationSession
+from opentrons.calibration.check import models as calibration_models
+from opentrons.calibration.check.session import CheckCalibrationSession,\
+    CalibrationSession
 from robot_server.service.models.session import SessionDetails
